requirements.txt
tldextract
argparse
colorama
dnspython